Resolved Conflicts:
	drivers/firmware/efivars.c
	fs/efivarsfs/file.c

+++ drivers/usb/c67x00/c67x00-ll-hpi.c
@@ -238,7 +238,7 @@ void c67x00_ll_hpi_disable_sofeop(struct c67x00_sie *sie)
 /* -------------------------------------------------------------------------- */
 /* Transactions */
 
-static inline u16 ll_recv_msg(struct c67x00_device *dev)
+static inline int ll_recv_msg(struct c67x00_device *dev)
 {
 	u16 res;
